# test/e2e/__snapshots__/test_openapi/test_OpenAPIの形が変わっていないことを確認.json
{
  "components": {
    "schemas": {
      "AccentPhrase": {
        "description": "アクセント句ごとの情報",
        "properties": {
          "accent": {
            "title": "アクセント箇所",
            "type": "integer"
          },
          "is_interrogative": {
            "default": false,
            "title": "疑問系かどうか",
            "type": "boolean"
          },
          "moras": {
            "items": {
              "$ref": "#/components/schemas/Mora"
            },
            "title": "モーラのリスト",
            "type": "array"
          },
          "pause_mora": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Mora"
              },
              {
                "type": "null"
              }
            ],
            "title": "後ろに無音を付けるかどうか"
          }
        },
        "required": [
          "moras",
          "accent"
        ],
        "title": "AccentPhrase",
        "type": "object"
      },
...
      "AivmManifest": {
        "description": "AIVM (Aivis Voice Model) マニフェストの定義",
        "properties": {
          "description": {
            "title": "音声合成モデルの説明 (省略時は空文字列になる)",
            "type": "string"
          },
          "manifest_version": {
            "title": "AIVM マニフェストのバージョン",
            "type": "string"
          },
          "model_architecture": {
            "title": "音声合成モデルのアーキテクチャ (音声合成技術の種類)",
            "type": "string"
          },
          "name": {
            "title": "音声合成モデルの名前",
            "type": "string"
          },
          "speakers": {
            "items": {
              "$ref": "#/components/schemas/AivmManifestSpeaker"
            },
            "title": "音声合成モデルに含まれる話者のリスト",
            "type": "array"
          },
          "uuid": {
            "title": "音声合成モデルの UUID",
            "type": "string"
          },
          "version": {
            "title": "音声合成モデルのバージョン",
            "type": "string"
          }
        },
        "required": [
          "manifest_version",
          "name",
          "description",
          "model_architecture",
          "uuid",
          "version",
          "speakers"
        ],
        "title": "AivmManifest",
        "type": "object"
      },
      "AivmManifestSpeaker": {
        "description": "AIVM (Aivis Voice Model) マニフェストの話者の定義\n画像やボイスサンプルは容量が大きいためマニフェストには含まれず、 別途ファイルとして AIVM に格納される",
        "properties": {
          "id": {
            "title": "話者の ID (この AIVM ファイル内で一意な 0 から始まる連番で、speaker_uuid とは異なる)",
            "type": "integer"
          },
          "name": {
            "title": "話者の名前",
            "type": "string"
          },
          "styles": {
            "items": {
              "$ref": "#/components/schemas/AivmManifestSpeakerStyle"
            },
            "title": "話者スタイルのリスト",
            "type": "array"
          },
          "supported_languages": {
            "items": {
              "type": "string"
            },
            "title": "話者の対応言語のリスト (ja, en, zh のような ISO 639-1 言語コード)",
            "type": "array"
          },
          "uuid": {
            "title": "話者の UUID (speaker_uuid と一致する)",
            "type": "string"
          },
          "version": {
            "title": "話者のバージョン",
            "type": "string"
          }
        },
        "required": [
          "name",
          "supported_languages",
          "id",
          "uuid",
          "version",
          "styles"
        ],
        "title": "AivmManifestSpeaker",
        "type": "object"
      },
      "AivmManifestSpeakerStyle": {
        "description": "AIVM (Aivis Voice Model) マニフェストの話者スタイルの定義",
        "properties": {
          "id": {
            "title": "スタイルの ID (この AIVM ファイル内で一意な 0 から始まる連番で、style_id とは異なる)",
            "type": "integer"
          },
          "name": {
            "title": "スタイルの名前",
            "type": "string"
          }
        },
        "required": [
          "name",
          "id"
        ],
        "title": "AivmManifestSpeakerStyle",
        "type": "object"
      },
      "AudioQuery": {
        "description": "音声合成用のクエリ",
        "properties": {
          "accent_phrases": {
            "items": {
              "$ref": "#/components/schemas/AccentPhrase"
            },
            "title": "アクセント句のリスト",
            "type": "array"
          },
          "intonationScale": {
            "description": "話者スタイルの声色の強弱を 0.0 ~ 2.0 の範囲で指定する (デフォルト: 1.0) 。\n値が大きいほどそのスタイルに近い抑揚がついた声になる。\n例えば話者スタイルが「うれしい」なら、値が大きいほどより嬉しそうな明るい話し方になる。\n一方スタイルによっては値を大きくしすぎると不自然な棒読みボイスになりがちなので、適宜調整が必要。\n全スタイルの平均であるノーマルスタイルには指定できない (値にかかわらず無視される) 。",
            "title": "全体のスタイルの強さ (「全体の抑揚」ではない点で VOICEVOX ENGINE と異なる)",
            "type": "number"
          },
          "kana": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "読み上げるテキストを指定する。\nVOICEVOX ENGINE では AquesTalk 風記法テキストが入る読み取り専用フィールドだが (音声合成時には無視される) 、AivisSpeech Engine では音声合成時に漢字や記号が含まれた通常の読み上げテキストも必要なため、苦肉の策で読み上げテキスト指定用のフィールドとして転用した。\nVOICEVOX ENGINE との互換性のため None や空文字列が指定された場合も動作するが、その場合はアクセント句から自動生成されたひらがな文字列が読み上げテキストになるため、不自然なイントネーションになってしまう。\n可能な限り kana に通常の読み上げテキストを指定した上で音声合成 API に渡すことを推奨する。",
            "title": "読み上げるテキスト (「読みの AquesTalk 風記法テキスト」ではない点で VOICEVOX ENGINE と異なる)"
          },
          "outputSamplingRate": {
            "title": "音声データの出力サンプリングレート",
            "type": "integer"
          },
          "outputStereo": {
            "title": "音声データをステレオ出力するか否か",
            "type": "boolean"
          },
          "pitchScale": {
            "description": "全体の音高を -0.15 ~ 0.15 の範囲で指定する (デフォルト: 0.0) 。\n値が大きいほど高い声になる。\nVOICEVOX ENGINE と異なり、この値を 0.0 から変更すると音質が劣化するため注意が必要。",
            "title": "全体の音高",
            "type": "number"
          },
          "postPhonemeLength": {
            "title": "音声の後の無音時間 (秒)",
            "type": "number"
          },
          "prePhonemeLength": {
            "title": "音声の前の無音時間 (秒)",
            "type": "number"
          },
          "speedScale": {
            "description": "全体の話速を 0.5 ~ 2.0 の範囲で指定する (デフォルト: 1.0) 。\n2.0 で 2 倍速、0.5 で 0.5 倍速になる。",
            "title": "全体の話速",
            "type": "number"
          },
          "tempoDynamicsScale": {
            "default": 1.0,
            "description": "話す速さの緩急の強弱を 0.0 ~ 2.0 の範囲で指定する (デフォルト: 1.0) 。\n値が大きいほどより早口で生っぽい抑揚がついた声になる。\nVOICEVOX ENGINE との互換性のため、未指定時はデフォルト値が適用される。",
            "title": "全体のテンポの緩急 (AivisSpeech Engine 固有のフィールド)",
            "type": "number"
          },
          "volumeScale": {
            "description": "全体の音量を 0.0 ~ 2.0 の範囲で指定する (デフォルト: 1.0) 。\n値が大きいほど大きな声になる。",
            "title": "全体の音量",
            "type": "number"
          }
        },
        "required": [
          "accent_phrases",
          "speedScale",
          "intonationScale",
          "pitchScale",
          "volumeScale",
          "prePhonemeLength",
          "postPhonemeLength",
          "outputSamplingRate",
          "outputStereo"
        ],
        "title": "AudioQuery",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "Mora": {
        "description": "モーラ（子音＋母音）ごとの情報",
        "properties": {
          "consonant": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "子音の音素"
          },
          "consonant_length": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "子音の音長"
          },
          "pitch": {
            "title": "音高",
            "type": "number"
          },
          "text": {
            "title": "文字",
            "type": "string"
          },
          "vowel": {
            "title": "母音の音素",
            "type": "string"
          },
          "vowel_length": {
            "title": "母音の音長",
            "type": "number"
          }
        },
        "required": [
          "text",
          "vowel",
          "vowel_length",
          "pitch"
        ],
        "title": "Mora",
        "type": "object"
      },
      "MorphableTargetInfo": {
        "properties": {
          "is_morphable": {
            "title": "指定した話者に対してモーフィングの可否",
            "type": "boolean"
          }
        },
        "required": [
          "is_morphable"
        ],
        "title": "MorphableTargetInfo",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "AivisSpeech の音声合成エンジンです。",
    "title": "AivisSpeech Engine",
    "version": "latest"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/aivm_models/{aivm_uuid}": {
      "delete": {
        "description": "音声合成モデルをアンインストールします。",
        "operationId": "uninstall_aivm_aivm_models__aivm_uuid__delete",
        "parameters": [
          {
            "description": "音声合成モデルの UUID",
            "in": "path",
            "name": "aivm_uuid",
            "required": true,
            "schema": {
              "description": "音声合成モデルの UUID",
              "title": "Aivm Uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Uninstall Aivm",
        "tags": [
          "音声合成モデル管理"
        ]
      },
      "post": {
        "description": "音声合成モデルをインストールします。\n音声合成モデルパッケージファイル (`.aivm`) をリクエストボディとして送信してください。",
        "operationId": "install_aivm_aivm_models__aivm_uuid__post",
        "parameters": [
          {
            "description": "音声合成モデルの UUID",
            "in": "path",
            "name": "aivm_uuid",
            "required": true,
            "schema": {
              "description": "音声合成モデルの UUID",
              "title": "Aivm Uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Install Aivm",
        "tags": [
          "音声合成モデル管理"
        ]
      }
    },
    "/aivm_models/{aivm_uuid}/manifest": {
      "get": {
        "description": "音声合成モデルの AIVM マニフェストを返します。",
        "operationId": "get_aivm_manifest_aivm_models__aivm_uuid__manifest_get",
        "parameters": [
          {
            "description": "音声合成モデルの UUID",
            "in": "path",
            "name": "aivm_uuid",
            "required": true,
            "schema": {
              "description": "音声合成モデルの UUID",
              "title": "Aivm Uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AivmManifest"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Aivm Manifest",
        "tags": [
          "音声合成モデル管理"
        ]
      }
    },
...
    "/morphable_targets": {
      "post": {
        "description": "指定されたベーススタイルに対してエンジン内の各話者がモーフィング機能を利用可能か返します。\nモーフィングの許可/禁止は `/speakers `の `speaker.supported_features.synthesis_morphing` に記載されています。\nプロパティが存在しない場合は、モーフィングが許可されているとみなします。\n返り値のスタイル ID は string 型なので注意。\nAivisSpeech Engine では話者ごとに発声タイミングが異なる関係で実装不可能なため (動作こそするが聴くに耐えない) 、\n全ての話者でモーフィングが禁止されています。",
        "operationId": "morphable_targets_morphable_targets_post",
        "parameters": [
          {
            "description": "AivisSpeech Engine ではサポートされていないパラメータです (常に無視されます) 。",
            "in": "query",
            "name": "core_version",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "AivisSpeech Engine ではサポートされていないパラメータです (常に無視されます) 。",
              "title": "Core Version"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "items": {
                  "type": "integer"
                },
                "title": "Base Style Ids",
                "type": "array"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "additionalProperties": {
                      "$ref": "#/components/schemas/MorphableTargetInfo"
                    },
                    "type": "object"
                  },
                  "title": "Response Morphable Targets Morphable Targets Post",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "指定したスタイルに対してエンジン内の話者がモーフィングが可能か判定する",
        "tags": [
          "音声合成"
        ]
      }
    },
...
    "/synthesis_morphing": {
      "post": {
        "description": "指定された 2 種類のスタイルで音声を合成、指定した割合でモーフィングした音声を得ます。\nモーフィングの割合は `morph_rate` で指定でき、0.0 でベースのスタイル、1.0 でターゲットのスタイルに近づきます。\nAivisSpeech Engine では話者ごとに発声タイミングが異なる関係で実装不可能なため (動作こそするが聴くに耐えない) 、\n常に 400 Bad Request を返します。",
        "operationId": "_synthesis_morphing_synthesis_morphing_post",
        "parameters": [
          {
            "in": "query",
            "name": "base_speaker",
            "required": true,
            "schema": {
              "title": "Base Speaker",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "target_speaker",
            "required": true,
            "schema": {
              "title": "Target Speaker",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "morph_rate",
            "required": true,
            "schema": {
              "maximum": 1.0,
              "minimum": 0.0,
              "title": "Morph Rate",
              "type": "number"
            }
          },
          {
            "description": "AivisSpeech Engine ではサポートされていないパラメータです (常に無視されます) 。",
            "in": "query",
            "name": "core_version",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "AivisSpeech Engine ではサポートされていないパラメータです (常に無視されます) 。",
              "title": "Core Version"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AudioQuery"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "audio/wav": {
                "schema": {
                  "format": "binary",
                  "type": "string"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "2種類のスタイルでモーフィングした音声を合成する",
        "tags": [
          "音声合成"
        ]
      }
    },
...
  }
}